import reanimated and surround it with try/catch
