+++ packages/expo-ui/src/swift-ui/State/optionalWorklets.ts
@@ -5,6 +5,8 @@
 let worklets: undefined | typeof import('react-native-worklets');
 try {
   worklets = require('react-native-worklets');
-} catch {}
+} catch (e) {
+  // Fail silently as worklet support is currently optional in Expo UI
+}
 
-export default worklets;
+export { worklets };
