+++ packages/expo-ui/src/swift-ui/State/utils.ts
@@ -1,0 +1,10 @@
+/**
+ * Extracts the native shared object ID from a SharedObject instance.
+ * Used internally to pass SharedObject references as view props.
+ */
+export function getStateId(state?: object): number | undefined {
+  if (!state) {
+    return undefined;
+  }
+  return (state as { __expo_shared_object_id__?: number }).__expo_shared_object_id__;
+}
