+++ packages/expo-ui/src/swift-ui/index.tsx
@@ -32,7 +32,7 @@ export * from './Slider';
 export * from './Spacer';
 export * from './Stepper';
 export * from './Text';
-export * from './State';
+export { useNativeState } from './State/useNativeState';
 export * from './SyncToggle';
 export * from './Toggle';
 export * from './TextField';
